"long long" causes compiler warnings when using libsodium from C++98 with -pedantic -Werror
- renamed "long long" to int64_t
- renamed "unsigned long long" to uint64_t
- include <stdint.h> in every header after <stddef.h>
- make check passes (on 64 bit Linux)

So you can't include them from a C++98 program with -pedantic -Werror

+++ src/libsodium/crypto_auth/hmacsha256/ref/hmac_hmacsha256.c
@@ -22,12 +22,12 @@ static const unsigned char iv[32] = {
   0x5b,0xe0,0xcd,0x19,
 } ;
 
-int crypto_auth(unsigned char *out,const unsigned char *in,unsigned long long inlen,const unsigned char *k)
+int crypto_auth(unsigned char *out,const unsigned char *in,uint64_t inlen,const unsigned char *k)
 {
   unsigned char h[32];
   unsigned char padded[128];
-  unsigned long long i;
-  unsigned long long bits = 512 + (inlen << 3);
+  uint64_t i;
+  uint64_t bits = 512 + (inlen << 3);
 
   for (i = 0;i < 32;++i) h[i] = iv[i];
