+++ src/libsodium/crypto_hash/sha512/ref/hash_sha512.c
@@ -20,14 +20,14 @@ static const unsigned char iv[64] = {
   0x5b,0xe0,0xcd,0x19,0x13,0x7e,0x21,0x79
 } ;
 
-typedef unsigned long long uint64;
+typedef uint64_t uint64;
 
-int crypto_hash(unsigned char *out,const unsigned char *in,unsigned long long inlen)
+int crypto_hash(unsigned char *out,const unsigned char *in,uint64_t inlen)
 {
   unsigned char h[64];
   unsigned char padded[256];
-  unsigned long long i;
-  unsigned long long bytes = inlen;
+  uint64_t i;
+  uint64_t bytes = inlen;
 
   for (i = 0;i < 64;++i) h[i] = iv[i];
