+++ src/libsodium/crypto_onetimeauth/crypto_onetimeauth.c
@@ -21,14 +21,14 @@ crypto_onetimeauth_primitive(void)
 
 int
 crypto_onetimeauth(unsigned char *out, const unsigned char *in,
-                   unsigned long long inlen, const unsigned char *k)
+                   uint64_t inlen, const unsigned char *k)
 {
     return crypto_onetimeauth_poly1305(out, in, inlen, k);
 }
 
 int
 crypto_onetimeauth_verify(const unsigned char *h, const unsigned char *in,
-                          unsigned long long inlen, const unsigned char *k)
+                          uint64_t inlen, const unsigned char *k)
 {
     return crypto_onetimeauth_poly1305_verify(h, in, inlen, k);
 }
