+++ src/libsodium/crypto_onetimeauth/poly1305/donna/auth_poly1305_donna.c
@@ -7,14 +7,14 @@
 
 int
 crypto_onetimeauth(unsigned char *out, const unsigned char *m,
-                   unsigned long long inlen, const unsigned char *key)
+                   uint64_t inlen, const unsigned char *key)
 {
 	uint32_t t0,t1,t2,t3;
 	uint32_t h0,h1,h2,h3,h4;
 	uint32_t r0,r1,r2,r3,r4;
 	uint32_t s1,s2,s3,s4;
 	uint32_t b, nb;
-	unsigned long long j;
+	uint64_t j;
 	uint64_t t[5];
 	uint64_t f0,f1,f2,f3;
 	uint32_t g0,g1,g2,g3,g4;
