+++ src/libsodium/crypto_stream/crypto_stream.c
@@ -20,7 +20,7 @@ crypto_stream_primitive(void)
 }
 
 int
-crypto_stream(unsigned char *c, unsigned long long clen,
+crypto_stream(unsigned char *c, uint64_t clen,
               const unsigned char *n, const unsigned char *k)
 {
     return crypto_stream_xsalsa20(c, clen, n, k);
@@ -29,7 +29,7 @@ crypto_stream(unsigned char *c, unsigned long long clen,
 
 int
 crypto_stream_xor(unsigned char *c, const unsigned char *m,
-                  unsigned long long mlen, const unsigned char *n,
+                  uint64_t mlen, const unsigned char *n,
                   const unsigned char *k)
 {
     return crypto_stream_xsalsa20_xor(c, m, mlen, n, k);
